# laravel-app/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/ratings": {
            "get": {
                "tags": [
                    "Ratings"
                ],
                "summary": "Obtenir la liste des évaluations de l'utilisateur connecté",
                "operationId": "getUserRatings",
                "responses": {
                    "200": {
                        "description": "Liste des évaluations récupérée avec succès"
                    },
                    "500": {
                        "description": "Erreur interne"
                    }
                }
            },
            "post": {
                "tags": [
                    "Ratings"
                ],
                "summary": "Créer ou mettre à jour une évaluation pour un plan",
                "operationId": "storeRating",
                "parameters": [
                    {
                        "name": "plan_id",
                        "in": "query",
                        "description": "ID du plan a noter",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    },
                    {
                        "name": "rating",
                        "in": "query",
                        "description": "Note du plan (1-5)",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "3"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Évaluation enregistrée avec succès"
                    },
                    "400": {
                        "description": "Données invalides"
                    },
                    "500": {
                        "description": "Erreur interne"
                    }
                }
            }
        },
        "/api/ratings/{id}": {
            "get": {
                "tags": [
                    "Ratings"
                ],
                "summary": "Afficher les détails d'une évaluation",
                "operationId": "getRating",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Évaluation récupérée avec succès"
                    },
                    "404": {
                        "description": "Évaluation non trouvée"
                    },
                    "500": {
                        "description": "Erreur interne"
                    }
                }
            },
            "put": {
                "tags": [
                    "Ratings"
                ],
                "summary": "Mettre à jour une évaluation existante",
                "operationId": "updateRating",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "plan_id",
                                    "rating"
                                ],
                                "properties": {
                                    "plan_id": {
                                        "description": "ID du plan évalué",
                                        "type": "integer"
                                    },
                                    "rating": {
                                        "description": "Note du plan (1-5)",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Évaluation mise à jour avec succès"
                    },
                    "400": {
                        "description": "Données invalides"
                    },
                    "404": {
                        "description": "Évaluation non trouvée"
                    },
                    "500": {
                        "description": "Erreur interne"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Ratings"
                ],
                "summary": "Supprimer une évaluation",
                "operationId": "deleteRating",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Évaluation supprimée avec succès"
                    },
                    "404": {
                        "description": "Évaluation non trouvée"
                    },
                    "500": {
                        "description": "Erreur interne"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}